+++ packages/purchasely/src/index.ts
@@ -485,5 +485,6 @@ export {
 } from './interceptor';
 export { PurchaselyBuilder } from './startBuilder';
 export { PLYPresentationView };
+export type { PLYPresentationViewResult } from './components/PLYPresentationView';
 
 export default Purchasely;